+++ lib/tescon/analyzer.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "rules/example_dsl"
+require_relative "rules/expect_be_nil"
 require_relative "rules/expect_eq"
 require_relative "rules/factory_bot"
 require_relative "rules/is_expected_eq"
@@ -24,6 +25,7 @@ module Tescon
       Rules::RspecDescribe.new,
       Rules::ExampleDsl.new,
       Rules::Subject.new,
+      Rules::ExpectBeNil.new,
       Rules::ExpectEq.new,
       Rules::IsExpectedEq.new
     ].freeze